refactor: move some eval functionality into split modules
This slightly reduces build parallelism, increasing compile times
slightly. However, this tradeoff is worth it for the better code
organisation.
