+++ primer/src/Primer/Eval.hs
@@ -1,5 +1,5 @@
 {-# LANGUAGE DuplicateRecordFields #-}
-{-# LANGUAGE ImpredicativeTypes #-}
+{-# LANGUAGE ViewPatterns #-}
 
 module Primer.Eval (
   -- The public API of this module
@@ -15,10 +15,9 @@ module Primer.Eval (
   LetRenameDetail (..),
   PushAppIntoLetrecDetail (..),
   ApplyPrimFunDetail (..),
+  -- Only exported for testing
   Locals,
   LocalLet (..),
-  tryPrimFun,
-  -- Only exported for testing
   tryReduceExpr,
   tryReduceType,
   findNodeByID,
@@ -28,27 +27,18 @@ module Primer.Eval (
 
 import Foreword
 
-import Control.Arrow ((***))
 import Control.Monad.Fresh (MonadFresh)
 import Data.Map.Strict qualified as Map
 import Data.Set qualified as Set
 import Optics (
-  Field1 (_1),
   elemOf,
-  filtered,
   getting,
-  mapping,
-  notElemOf,
-  set,
-  to,
   view,
   (%),
   (^.),
-  (^..),
   _2,
  )
 import Primer.Core (
-  Bind' (..),
   CaseBranch' (..),
   Expr,
   Expr' (..),
@@ -57,37 +47,21 @@ import Primer.Core (
   ID,
   LVarName,
   LocalName (LocalName, unLocalName),
-  Meta,
   TmVarRef (..),
   TyVarName,
   Type,
   Type' (..),
-  TypeCache,
   bindName,
-  getID,
-  _exprMetaLens,
-  _type,
  )
-import Primer.Core.DSL (ann, hole, letType, let_, tEmptyHole)
-import Primer.Core.Transform (removeAnn, renameLocalVar, renameTyVarExpr, unfoldAPP, unfoldApp)
+import Primer.Core.Transform (removeAnn, unfoldAPP, unfoldApp)
 import Primer.Core.Utils (
-  concreteTy,
-  forgetMetadata,
   freeVars,
   freeVarsTy,
-  regenerateExprIDs,
   regenerateTypeIDs,
   _freeTmVars,
   _freeTyVars,
-  _freeVars,
-  _freeVarsTy,
  )
-import Primer.Def (
-  ASTDef (..),
-  Def (..),
-  DefMap,
-  defPrim,
- )
+import Primer.Def (DefMap)
 import Primer.Eval.Detail (
   ApplyPrimFunDetail (..),
   BetaReductionDetail (..),
@@ -96,13 +70,25 @@ import Primer.Eval.Detail (
   GlobalVarInlineDetail (..),
   LetRemovalDetail (..),
   LetRenameDetail (..),
+  LocalLet (LLet, LLetRec, LLetType),
   LocalVarInlineDetail (..),
+  Locals,
   PushAppIntoLetrecDetail (..),
+  RHSCaptured (Capture, NoCapture),
+  tryCaseReduction,
+  tryInlineGlobal,
+  tryInlineLocal,
+  tryLetRemoval,
+  tryPrimFun,
+  tryReduceBETA,
+  tryReduceBeta,
+  tryReducePrim,
+  tryReducePush,
  )
 import Primer.Eval.EvalError (EvalError (..))
-import Primer.Name (Name, unName, unsafeMkName)
+import Primer.Name (Name)
 import Primer.Name.Fresh (isFresh, isFreshTy)
-import Primer.Primitives (PrimDef, primFunDef)
+import Primer.Primitives.PrimDef (PrimDef)
 import Primer.Zipper (
   ExprZ,
   FoldAbove,
@@ -119,24 +105,6 @@ import Primer.Zipper (
   unfocusExpr,
   unfocusType,
  )
-
--- | A map from local variable names to the ID of their binding, their bound
--- value and whether anything in their value would be captured by an intervening
--- binder. (NB: a non-recursive let can "capture itself" and this is also
--- detected here.)
--- Since each entry must have a value, this only includes let(rec) bindings.
--- Lambda bindings must be reduced to a let before their variables can appear here.
---
--- Values of this type are constructed by 'findNodeByID'.
-type Locals = Map Name (ID, LocalLet, RHSCaptured)
-
-data LocalLet = LLet Expr | LLetRec Expr | LLetType Type
-  deriving (Eq, Show)
-
-data RHSCaptured
-  = NoCapture
-  | Capture
-  deriving (Eq, Show)
 
 -- | Perform one step of reduction on the node with the given ID
 -- Returns the new expression and its redexes.
@@ -358,67 +326,8 @@ redexes primDefs = go mempty
     insertTy :: TyVarName -> Type -> Map TyVarName (Set TyVarName) -> Map TyVarName (Set TyVarName)
     insertTy x t = Map.insert x (freeVarsTy t) . removeTy x
 
--- | Extract the cached type information from the metadata of an AST node.
-annOf :: Meta a -> a
-annOf = view _type
-
--- | Set the cached type information of the root node of the given expression to the given value.
-annotate :: Maybe TypeCache -> Expr -> Expr
-annotate = set (_exprMetaLens % _type)
-
--- | This function helps us create let bindings which are easy to substitute
--- without causing variable capture.
--- It takes as arguments:
--- - a variable name, @x@ (the variable we would like to bind)
--- - some names to avoid @vs@ (normally the free (type and term) variables in
---   the term we would like to bind)
--- - the term @t[x]@ we would like the binding to scope over (in which which
---   @x@ presumably appears free)
--- It will then modify the original name until it finds one that:
--- - doesn't clash with any of the @vs@
--- - can be safely used instead of the original name in the lambda body
--- Thus, it will return @y@ and @t[y]@ for some @y@ distinct from each
--- of the @vs@, and such that the renaming is trivial (i.e. we do not need to
--- alpha convert any binders in @t@ to avoid capture).
---
--- We assume that the original name is safe to use, so we return it
--- unchanged if it doesn't clash with a free variable in the argument.
---
--- The reason this eases future substitution is that we avoid making terms such
--- as @let x = C x in D x x@ where you cannot inline just one occurrence of the
--- @x@ without causing capture.
---
--- See 'Tests.Eval.unit_tryReduce_beta_name_clash' for an example of where this is useful.
-makeSafeLetBinding :: LVarName -> Set Name -> Expr -> (LVarName, Expr)
-makeSafeLetBinding = makeSafeLetBinding' renameLocalVar
-
--- | As 'makeSafeLetBinding', but for Λ applications
-makeSafeLetTypeBinding :: TyVarName -> Set Name -> Expr -> (TyVarName, Expr)
-makeSafeLetTypeBinding = makeSafeLetBinding' renameTyVarExpr
-
--- Helper for makeSafeLet{,Type}Binding
-makeSafeLetBinding' ::
-  (LocalName k -> LocalName k -> Expr -> Maybe Expr) ->
-  LocalName k ->
-  Set Name ->
-  Expr ->
-  (LocalName k, Expr)
-makeSafeLetBinding' _ name others body | Set.notMember (unLocalName name) others = (name, body)
-makeSafeLetBinding' rename name others body = go 0
-  where
-    go :: Int -> (LocalName k, Expr)
-    go n =
-      let newName' = unsafeMkName $ unName (unLocalName name) <> show n
-          newName = LocalName newName'
-       in if Set.member newName' others
-            then go (n + 1)
-            else case rename name newName body of
-              Just body' -> (newName, body')
-              Nothing -> go (n + 1)
-
 -- | Given a context of local and global variables and an expression, try to reduce that expression.
 -- Expects that the expression is redex and will throw an error if not.
--- TODO: consider using view patterns for these cases
 tryReduceExpr ::
   (MonadFresh ID m, MonadError EvalError m) =>
   DefMap ->
@@ -426,99 +335,10 @@ tryReduceExpr ::
   Expr ->
   m (Expr, EvalDetail)
 tryReduceExpr globals locals = \case
-  -- Beta reduction (no annotation)
-  -- (\x. e1) e2 ==> let x = e2 in e1
-  App mApp lam@(Lam _ x body) arg -> do
-    let (x', body') = makeSafeLetBinding x (freeVars arg) body
-    expr <- annotate (annOf mApp) <$> let_ x' (pure arg) (pure body')
-    pure
-      ( expr
-      , BetaReduction
-          BetaReductionDetail
-            { before = App mApp lam arg
-            , after = expr
-            , bindingName = x
-            , lambdaID = lam ^. _id
-            , letID = expr ^. _id
-            , argID = arg ^. _id
-            , bodyID = body ^. _id
-            , types = Nothing
-            }
-      )
-  -- Beta reduction (with annotation)
-  -- (\x. e1 : A -> B) e2 ==> let x = e2 : A in e1 : B
-  App mApp annotation@(Ann _ lam@(Lam _ x body) ty) arg -> do
-    let (x', body') = makeSafeLetBinding x (freeVars arg) body
-        -- The annotation is a hole. This means we can't trust that the lambda has the right type for
-        -- this context. Specifically:
-        -- - the argument may not have the right type for the lambda body
-        -- - the lambda body may not have the right type for the application
-        -- To deal with this, we put both the argument and the body in holes.
-        -- TODO: explain this in the detail view.
-        holeAnn = do
-          lty <- tEmptyHole
-          rty <- tEmptyHole
-          l <- let_ x' (hole (pure arg)) (pure body')
-          e <- annotate (annOf mApp) <$> hole (pure l)
-          pure (e, l, Just (lty, rty))
-    (expr, letexpr, types) <- case ty of
-      -- The annotation is a function type, as expected
-      (TFun _ lty rty) -> do
-        l <- let_ x' (ann (pure arg) (pure lty)) (pure body')
-        e <- annotate (annOf mApp) <$> ann (pure l) (pure rty)
-        pure (e, l, Just (lty, rty))
-      TEmptyHole _ -> holeAnn
-      THole _ _ -> holeAnn
-      -- The annotation is of some other form, which we can't handle
-      _ -> throwError $ BadLambdaAnnotation annotation
-    pure
-      ( expr
-      , BetaReduction
-          BetaReductionDetail
-            { before = App mApp annotation arg
-            , after = expr
-            , bindingName = x
-            , lambdaID = lam ^. _id
-            , letID = letexpr ^. _id
-            , argID = arg ^. _id
-            , bodyID = body ^. _id
-            , types = types
-            }
-      )
-  -- (letrec x : T = t in λ ...) e  ~> letrec x : T = t in ((λ...) e)
-  before@(App mApp (Letrec mLet x e1 t lam@Lam{}) e2) | isFresh x e2 -> do
-    -- We push the application into the letrec, in order to enable it to reduce in a subsequent
-    -- step. This does not cause capture, as we have checked that x is not free in e2.
-    let expr = annotate (annOf mApp) $ Letrec mLet x e1 t (App mApp lam e2)
-    pure
-      ( expr
-      , PushAppIntoLetrec
-          PushAppIntoLetrecDetail
-            { before = before
-            , after = expr
-            , argID = e2 ^. _id
-            , letrecID = mLet ^. _id
-            , lamID = lam ^. _id
-            , letBindingName = x
-            , isTypeApplication = False
-            }
-      )
-
-  -- apply primitive function
-  before@App{}
-    | Just (name, args, e) <- tryPrimFun (Map.mapMaybe defPrim globals) before -> do
-        expr <- e
-        pure
-          ( expr
-          , ApplyPrimFun
-              ApplyPrimFunDetail
-                { before = before
-                , after = expr
-                , name = name
-                , argIDs = args ^. mapping _id
-                }
-          )
-
+  (tryReduceBeta -> Just m) -> second BetaReduction <$> m
+  (tryReduceBETA -> Just m) -> second BETAReduction <$> m
+  (tryReducePush -> Just m) -> second PushAppIntoLetrec <$> m
+  (tryReducePrim globals -> Just m) -> second ApplyPrimFun <$> m
   -- Beta reduction of an inner application
   -- This rule is theoretically redundant but because we render nested applications with just one
   -- 'App' node in the tupled style, the user can only select the top-most application even if the
@@ -531,73 +351,7 @@ tryReduceExpr globals locals = \case
     -- Try to reduce e1
     (e1', detail) <- tryReduceExpr globals locals e1
     pure (App mApp e1' e2, detail)
-  -- Beta reduction of big lambda (no annotation)
-  -- (Λx. e) t ==> let type x = t in e
-  APP mAPP lam@(LAM _ x body) arg -> do
-    let (x', body') = makeSafeLetTypeBinding x (Set.map unLocalName $ freeVarsTy arg) body
-    expr <- annotate (annOf mAPP) <$> letType x' (pure arg) (pure body')
-    pure
-      ( expr
-      , BETAReduction
-          BetaReductionDetail
-            { before = APP mAPP lam arg
-            , after = expr
-            , bindingName = x
-            , lambdaID = lam ^. _id
-            , letID = expr ^. _id
-            , argID = arg ^. _id
-            , bodyID = body ^. _id
-            , types = Nothing
-            }
-      )
-  -- Beta reduction of big lambda (with annotation)
-  -- With the current editor K is always KType, so the annotation is a
-  -- bit pointless, but we include this rule for completeness.
-  -- This is what we technically should do:
-  --   (Λx. e : ∀a : K. B) t ==> let type x = t in e : [t/a]B
-  -- But performing the substitution [t/a]B is a bit of a pain when you have to ensure ID uniqueness
-  -- and worry about other metadata, so for simplicity we just drop the annotation.
-  -- This might change in future if we decide we want to keep it.
-  -- So this is what we actually do:
-  --   (Λx. e : ∀a : K. B) t ==> let type x = t in e
-  APP mAPP annotation@(Ann _ lam@(LAM _ x body) ty) t -> do
-    let (x', body') = makeSafeLetTypeBinding x (Set.map unLocalName $ freeVarsTy t) body
-    case ty of
-      (TForall _ _ k b) -> do
-        expr <- annotate (annOf mAPP) <$> letType x' (pure t) (pure body')
-        pure
-          ( expr
-          , BETAReduction
-              BetaReductionDetail
-                { before = APP mAPP annotation t
-                , after = expr
-                , bindingName = x
-                , lambdaID = lam ^. _id
-                , letID = expr ^. _id
-                , argID = t ^. _id
-                , bodyID = body ^. _id
-                , types = Just (k, b)
-                }
-          )
-      _ -> throwError $ BadBigLambdaAnnotation annotation
-  -- (letrec x : T = t in Λ ...) e  ~>  letrec x : T = t in ((Λ ...) e)
-  before@(APP mApp (Letrec mLet x e1 t lam@LAM{}) e2) | isFreshTy x e2 -> do
-    -- We push the application into the letrec, in order to enable it to reduce in a subsequent
-    -- step. This does not cause capture, as we have checked that x is not free in e2.
-    let expr = annotate (annOf mApp) $ Letrec mLet x e1 t (APP mApp lam e2)
-    pure
-      ( expr
-      , PushAppIntoLetrec
-          PushAppIntoLetrecDetail
-            { before = before
-            , after = expr
-            , argID = e2 ^. _id
-            , letrecID = mLet ^. _id
-            , lamID = lam ^. _id
-            , letBindingName = x
-            , isTypeApplication = True
-            }
-      )
+
   -- Beta reduction of an inner big lambda application
   -- This rule is theoretically redundant but because we render nested applications with just one
   -- 'APP' node in the tupled style, the user can only select the top-most application even if the
@@ -610,146 +364,11 @@ tryReduceExpr globals locals = \case
     -- Try to reduce e
     (e', detail) <- tryReduceExpr globals locals e
     pure (APP mAPP e' t, detail)
-  -- Inline local variable
-  -- x=e |- x ==> e
-  -- If the variable is not in the local set, that's fine - it just means it is bound by a lambda
-  -- that hasn't yet been reduced.
-  Var mVar (LocalVarRef x)
-    | Just (i, l, NoCapture) <- Map.lookup (unLocalName x) locals -> do
-        e <- case l of
-          LLet e' -> pure e'
-          LLetRec e' -> pure e'
-          LLetType _ -> throwError NotRedex
-        -- Since we're duplicating @e@, we must regenerate all its IDs.
-        e' <- regenerateExprIDs e
-        pure
-          ( e'
-          , LocalVarInline
-              LocalVarInlineDetail
-                { letID = i
-                , varID = mVar ^. _id
-                , valueID = e ^. _id
-                , bindingName = x
-                , replacementID = e' ^. _id
-                , isTypeVar = False
-                }
-          )
-  -- Inline global variable
-  -- (f = e : t) |- f ==> e : t
-  Var mVar (GlobalVarRef x) | Just (DefAST def) <- Map.lookup x globals -> do
-    -- Since we're duplicating the definition, we must regenerate all its IDs.
-    e <- regenerateExprIDs (astDefExpr def)
-    t <- regenerateTypeIDs (astDefType def)
-    expr <- ann (pure e) (pure t)
-    pure
-      ( expr
-      , GlobalVarInline
-          GlobalVarInlineDetail
-            { var = Var mVar (GlobalVarRef x)
-            , def = def
-            , after = expr
-            }
-      )
-  expr@(Let meta x e body)
-    -- Redundant let removal
-    -- let x = e1 in e2 ==> e2    if x not free in e2
-    | notElemOf (getting _freeTmVars % _2) x body -> mkLetRemovalDetail expr body x meta
-    -- Renaming a potentially self-capturing let
-    -- let x = f[x] in g[x] ==> let y = f[x] in g[y]
-    | otherwise -> mkLetRenameDetail expr body (Left (x, e)) meta
-  expr@(Letrec meta x _ _ body)
-    -- Redundant letrec removal
-    -- letrec x = e in e2 ==> e2  if x not free in e2
-    | notElemOf (getting _freeTmVars % _2) x body -> mkLetRemovalDetail expr body x meta
-  expr@(LetType meta x t body)
-    -- Redundant letType removal
-    -- let type x = t in e ==> e  if x not free in e
-    | notElemOf (getting _freeTyVars % _2) x body -> mkLetRemovalDetail expr body x meta
-    -- Renaming a potentially self-capturing letType
-    -- let type x = f[x] in g[x] ==> let type y = f[x] in g[y]
-    | otherwise -> mkLetRenameDetail expr body (Right (x, t)) meta
-  -- Case reduction
-  -- If the scrutinee starts with a constructor, we can reduce the case expression.
-  -- We do that by picking the branch with the same constructor.
-  -- For each variable bound by the branch pattern, we create a let with the same name, binding the
-  -- corresponding argument in the scrutinee.
-  Case m scrut branches
-    | (expr, termArgs) <- unfoldApp (removeAnn scrut)
-    , (Con mCon c, _typeArgs) <- unfoldAPP expr ->
-        do
-          -- Find the branch with the same constructor
-          case find (\(CaseBranch n _ _) -> n == c) branches of
-            Just (CaseBranch _ binds rhs) -> do
-              -- Check that we have as many term args as bindings
-              when (length binds /= length termArgs) $ throwError CaseBranchBindingLengthMismatch
-              -- We need to rename bindings to avoid variable capture.
-              -- See Note [Case reduction and variable capture]
-              -- in EvalFull
-              let (rhs', binds') = mapAccumR (\r (Bind _ x) -> swap $ makeSafeLetBinding x (foldMap freeVars termArgs) r) rhs binds
-              -- Construct a let for each bind
-              let makeLet (x, e) rest = let_ x (pure e) (pure rest)
-              (expr', letIDs) <-
-                foldrM
-                  ( \a (e, lets) -> do
-                      l <- makeLet a e
-                      pure (l, l ^. _id : lets)
-                  )
-                  (rhs', [])
-                  (zip binds' termArgs)
-              pure
-                ( expr'
-                , CaseReduction
-                    CaseReductionDetail
-                      { before = Case m scrut branches
-                      , after = expr'
-                      , targetID = scrut ^. _id
-                      , targetCtorID = mCon ^. _id
-                      , ctorName = c
-                      , targetArgIDs = map (^. _id) termArgs
-                      , branchBindingIDs = map (^. _id) binds
-                      , branchRhsID = rhs ^. _id
-                      , letIDs = letIDs
-                      }
-                )
-            Nothing -> throwError NoMatchingCaseBranch
+  (tryInlineLocal locals -> Just m) -> second LocalVarInline <$> m
+  (tryInlineGlobal globals -> Just m) -> second GlobalVarInline <$> m
+  (tryLetRemoval -> Just m) -> second (either LetRename LetRemoval) <$> m
+  (tryCaseReduction -> Just m) -> second CaseReduction <$> m
   _ -> throwError NotRedex
-  where
-    mkLetRemovalDetail expr body x meta =
-      pure
-        ( body
-        , LetRemoval
-            LetRemovalDetail
-              { before = expr
-              , after = body
-              , bindingName = unLocalName x
-              , letID = meta ^. _id
-              , bodyID = body ^. _id
-              }
-        )
-    mkLetRenameDetail expr body binding meta = do
-      (x, y, occ, expr') <- case binding of
-        Left (x, e) -> do
-          let (y, body') = makeSafeLetBinding x (freeVars e) body
-          let idName = either (getID *** unLocalName) (getID *** unLocalName)
-          let occ = e ^.. _freeVars % to idName % filtered ((== unLocalName x) . snd) % _1
-          (unLocalName x,unLocalName y,occ,) <$> let_ y (pure e) (pure body')
-        Right (x, ty) -> do
-          let (y, body') = makeSafeLetTypeBinding x (Set.map unLocalName $ freeVarsTy ty) body
-          let occ = ty ^.. getting _freeVarsTy % to (first getID) % filtered ((== x) . snd) % _1
-          (unLocalName x,unLocalName y,occ,) <$> letType y (pure ty) (pure body')
-      pure
-        ( expr'
-        , LetRename
-            LetRenameDetail
-              { before = expr
-              , after = expr'
-              , bindingNameOld = x
-              , bindingNameNew = y
-              , letID = meta ^. _id
-              , bindingOccurrences = occ
-              , bodyID = body ^. _id
-              }
-        )
 
 tryReduceType ::
   (MonadFresh ID m, MonadError EvalError m) =>
@@ -789,24 +408,3 @@ munless x b = if b then mempty else x
 -- It's like 'Control.Monad.when' but for Monoids rather than Applicatives.
 mwhen :: Monoid a => a -> Bool -> a
 mwhen x b = if b then x else mempty
-
--- | If this node is a reducible application of a primitive, return the name of the primitive, the arguments, and
--- (a computation for building) the result.
-tryPrimFun :: Map GVarName PrimDef -> Expr -> Maybe (GVarName, [Expr], forall m. MonadFresh ID m => m Expr)
-tryPrimFun primDefs expr
-  | -- Since no primitive functions are polymorphic, there is no need to unfoldAPP
-    (Var _ (GlobalVarRef name), args) <- bimap stripAnns (map stripAnns) $ unfoldApp expr
-  , Just x <- Map.lookup name primDefs
-  , Right e <- primFunDef x $ forgetMetadata <$> args =
-      Just (name, args, e)
-  | otherwise = Nothing
-  where
-    -- We have to be able to apply a primitive in the presence of type annotations.
-    -- This is important because other evaluation steps may introduce unnecessary annotations,
-    -- so we need to be able to ignore them (as we also do in the case of beta reduction).
-    -- During evaluation, we may choose to hide annotations anyway, so they really shouldn't make a difference to
-    -- what can be evaluated.
-    -- Note that it's only safe to remove concrete annotations, since holes can act as type-changing casts.
-    stripAnns = \case
-      Ann _ e t | concreteTy t -> stripAnns e
-      e -> e
