+++ primer/src/Primer/Eval/Let.hs
@@ -1,13 +1,24 @@
 {-# LANGUAGE DuplicateRecordFields #-}
 
-module Primer.Eval.Let (LetRemovalDetail (..), LetRenameDetail (..)) where
+module Primer.Eval.Let (LetRemovalDetail (..), LetRenameDetail (..), tryLetRemoval) where
 
 import Foreword
 
+import Control.Arrow ((***))
+import Control.Monad.Fresh (MonadFresh)
+import Data.Set qualified as Set
+import Optics (filtered, getting, notElemOf, to, (%), (^.), (^..), _1, _2)
 import Primer.Core (
   Expr,
+  Expr' (Let, LetType, Letrec),
   ID,
+  LocalName (unLocalName),
+  getID,
+  _id,
  )
+import Primer.Core.DSL (letType, let_)
+import Primer.Core.Utils (freeVars, freeVarsTy, _freeTmVars, _freeTyVars, _freeVars, _freeVarsTy)
+import Primer.Eval.Utils (makeSafeLetBinding, makeSafeLetTypeBinding)
 import Primer.JSON (CustomJSON (CustomJSON), FromJSON, PrimerJSON, ToJSON)
 import Primer.Name (Name)
 
@@ -44,3 +55,66 @@ data LetRenameDetail = LetRenameDetail
   }
   deriving (Eq, Show, Generic)
   deriving (FromJSON, ToJSON) via PrimerJSON LetRenameDetail
+
+tryLetRemoval ::
+  MonadFresh ID m =>
+  Expr ->
+  Maybe (m (Expr, Either LetRenameDetail LetRemovalDetail))
+tryLetRemoval = \case
+  expr@(Let meta x e body)
+    -- Redundant let removal
+    -- let x = e1 in e2 ==> e2    if x not free in e2
+    | notElemOf (getting _freeTmVars % _2) x body -> mkLetRemovalDetail expr body x meta
+    -- Renaming a potentially self-capturing let
+    -- let x = f[x] in g[x] ==> let y = f[x] in g[y]
+    | otherwise -> mkLetRenameDetail expr body (Left (x, e)) meta
+  expr@(Letrec meta x _ _ body)
+    -- Redundant letrec removal
+    -- letrec x = e in e2 ==> e2  if x not free in e2
+    | notElemOf (getting _freeTmVars % _2) x body -> mkLetRemovalDetail expr body x meta
+  expr@(LetType meta x t body)
+    -- Redundant letType removal
+    -- let type x = t in e ==> e  if x not free in e
+    | notElemOf (getting _freeTyVars % _2) x body -> mkLetRemovalDetail expr body x meta
+    -- Renaming a potentially self-capturing letType
+    -- let type x = f[x] in g[x] ==> let type y = f[x] in g[y]
+    | otherwise -> mkLetRenameDetail expr body (Right (x, t)) meta
+  _ -> Nothing
+  where
+    mkLetRemovalDetail expr body x meta =
+      Just $
+        pure
+          ( body
+          , Right $
+              LetRemovalDetail
+                { before = expr
+                , after = body
+                , bindingName = unLocalName x
+                , letID = meta ^. _id
+                , bodyID = body ^. _id
+                }
+          )
+    mkLetRenameDetail expr body binding meta = Just $ do
+      (x, y, occ, expr') <- case binding of
+        Left (x, e) -> do
+          let (y, body') = makeSafeLetBinding x (freeVars e) body
+          let idName = either (getID *** unLocalName) (getID *** unLocalName)
+          let occ = e ^.. _freeVars % to idName % filtered ((== unLocalName x) . snd) % _1
+          (unLocalName x,unLocalName y,occ,) <$> let_ y (pure e) (pure body')
+        Right (x, ty) -> do
+          let (y, body') = makeSafeLetTypeBinding x (Set.map unLocalName $ freeVarsTy ty) body
+          let occ = ty ^.. getting _freeVarsTy % to (first getID) % filtered ((== x) . snd) % _1
+          (unLocalName x,unLocalName y,occ,) <$> letType y (pure ty) (pure body')
+      pure
+        ( expr'
+        , Left $
+            LetRenameDetail
+              { before = expr
+              , after = expr'
+              , bindingNameOld = x
+              , bindingNameNew = y
+              , letID = meta ^. _id
+              , bindingOccurrences = occ
+              , bodyID = body ^. _id
+              }
+        )
